stockagent_rt/requirements.txt
# Python 3.13 compatible dependencies for StockAgent RT demo
pandas>=2.2.2,<2.3
numpy>=1.26.4,<3.0.0
python-dateutil>=2.8.2,<3
requests>=2.32.3,<3
matplotlib>=3.9.0,<3.10
seaborn>=0.13.2,<0.14
statsmodels>=0.14.2,<0.15
scikit-learn>=1.5.1,<1.6
pydantic>=2.8.2,<3
python-dotenv>=1.0.1,<2
tenacity>=9.0.0,<10
# LLM clients
openai>=1.40.0,<2
google-generativeai>=0.7.2,<0.8
rich>=13.7.1,<14
